% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagEmaPrediction.R
\docType{package}
\name{CovCoagEmaPrediction}
\alias{CovCoagEmaPrediction}
\title{CovCoagEmaPrediction}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
